fix: resolve issues caused by folder rename
