+++ .husky/pre-commit
@@ -1,0 +1,1 @@
+npm run format && npm run type-check
