+++ src/shared/ui/card/ReviewWriting.tsx
@@ -3,7 +3,7 @@
 import Image from 'next/image';
 import { MovingTypeChip } from '@/shared/ui/chip';
 import { StarSelect } from '@/shared/ui/star';
-import Button from '@/shared/ui/Button/Button';
+import { Button } from '@/shared/ui/button';
 import { useState } from 'react';
 
 interface ReviewWritingProps {
